Expanded Application User model, Implemented User Profile Page and Edit functionality to change the user profile details
Added social media and more info to the Application User model. Scaffolded and expand Registration. Added default avatar and profile photos that are set after registration and before user upload his own. Implemented User Profile Details Page with the user`s info and the list of spots that are uploaded by the user. Page and Edit functionality with option to change the profile picture.